pyproject.toml
[tool.poetry]
name = "api-for-library-management"
version = "0.1.0"
description = ""
authors = ["Никита Шидогубов <nikita.sh@gklifts.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
coreapi = "^2.3.3"
setuptools = "^75.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
